TrackPaginator: Replace with more autonomous version
The updated TrackPaginator nolonger directly communicates with the Tracks class. Instead it takes an array of tracks. While this is less database efficent, it makes the TrackPaginator more versatile, allowing it to be used by other pages that may not need to display every track in the database.

This also deprecates the 'after' parameter in Tracks::getAll().

NOTE: This commit does not update pages/tracks.php - so the paginator bar will disappear in this commit.